agri-main/ml-service/requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
Flask # We'll use Flask to create a simple API